Fix Mac dev setup issues [#5442].
A trailing slash in devtools/compile_translations.sh caused that script to fail
because the parameter to pybabel's -l option ended up blank instead of
containing the language code.

A few files and directories were missing from .gitignore and hence wound up
in the list of local changes to be committed.