Don't disable existing loggers when initialising Celery.
Prior to this change, the Celery logging did not include any of the logging
calls within MediaGoblin because `disable_existing_loggers` defaulted to True.
This was unhelpful and inconsistent with the logging behaviour of the web process.

The MediaGoblin web process sets logging output based on the configuration in
paste.ini. This is loaded by the `paster` program, rather than MediaGoblin
itself.

The MediaGoblin celery process manually loads its logging config from paste.ini,
but previously defaulted to `disable_existing_loggers=True`, meaning that none
of the application logging flowed through unless the logger was explicitly added
to paste.ini.